src/backend/requirements.txt
asgiref==3.9.1
Django==5.2.6
sqlparse==0.5.3
tzdata==2025.2
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==42.0.8
django-cors-headers==4.7.0
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.1.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
